Move extensions to api so they are public avalible
