+++ src/main/kotlin/org/ender_development/catalyx/api/v1/extensions/Any.kt
@@ -1,4 +1,4 @@
-package org.ender_development.catalyx.core.utils.extensions
+package org.ender_development.catalyx.api.v1.extensions
 
 import org.ender_development.catalyx.core.utils.validation.IValidator
 import org.ender_development.catalyx.core.utils.validation.ValidationBuilder
