+++ src/main/kotlin/org/ender_development/catalyx/api/v1/extensions/EnumDyeColor.kt
@@ -1,4 +1,4 @@
-package org.ender_development.catalyx.core.utils.extensions
+package org.ender_development.catalyx.api.v1.extensions
 
 import net.minecraft.item.EnumDyeColor
 import org.ender_development.catalyx.core.utils.ColorMapping
