+++ src/main/kotlin/org/ender_development/catalyx/api/v1/extensions/Logger.kt
@@ -1,6 +1,6 @@
 @file:Suppress("NOTHING_TO_INLINE")
 
-package org.ender_development.catalyx.core.utils.extensions
+package org.ender_development.catalyx.api.v1.extensions
 
 import org.apache.logging.log4j.LogManager
 import org.apache.logging.log4j.Logger
