+++ src/main/kotlin/org/ender_development/catalyx/api/v1/extensions/NBTTagCompound.kt
@@ -1,4 +1,4 @@
-package org.ender_development.catalyx.core.utils.extensions
+package org.ender_development.catalyx.api.v1.extensions
 
 import net.minecraft.nbt.NBTTagCompound
 import net.minecraft.nbt.NBTTagLongArray
