+++ src/main/kotlin/org/ender_development/catalyx/api/v1/extensions/Set.kt
@@ -1,4 +1,4 @@
-package org.ender_development.catalyx.core.utils.extensions
+package org.ender_development.catalyx.api.v1.extensions
 
 import it.unimi.dsi.fastutil.objects.ObjectSets
 import java.util.*
