+++ src/main/kotlin/org/ender_development/catalyx/core/blocks/multiblock/CenterBlock.kt
@@ -8,9 +8,9 @@ import net.minecraft.tileentity.TileEntity
 import net.minecraft.util.math.AxisAlignedBB
 import net.minecraft.util.math.BlockPos
 import net.minecraft.world.World
+import org.ender_development.catalyx.api.v1.extensions.getHorizontalSurroundings
 import org.ender_development.catalyx.core.ICatalyxMod
 import org.ender_development.catalyx.core.blocks.BaseRotatableTileBlock
-import org.ender_development.catalyx.core.utils.extensions.getHorizontalSurroundings
 
 open class CenterBlock<T>(mod: ICatalyxMod, name: String, tileClass: Class<T>, guiId: Int, vararg components: IMultiblockEdge) : BaseRotatableTileBlock(
 	mod, name, tileClass, guiId
