+++ src/main/kotlin/org/ender_development/catalyx/core/network/ButtonPacket.kt
@@ -8,10 +8,10 @@ import net.minecraftforge.fml.common.network.simpleimpl.IMessage
 import net.minecraftforge.fml.common.network.simpleimpl.IMessageHandler
 import net.minecraftforge.fml.common.network.simpleimpl.MessageContext
 import org.ender_development.catalyx.Catalyx
+import org.ender_development.catalyx.api.v1.extensions.readString
+import org.ender_development.catalyx.api.v1.extensions.writeString
 import org.ender_development.catalyx.core.client.button.AbstractButtonWrapper
 import org.ender_development.catalyx.core.tiles.helper.IButtonTile
-import org.ender_development.catalyx.core.utils.extensions.readString
-import org.ender_development.catalyx.core.utils.extensions.writeString
 
 class ButtonPacket() : IMessage {
 	private lateinit var blockPos: BlockPos
