+++ src/main/kotlin/org/ender_development/catalyx/core/recipes/Recipe.kt
@@ -7,13 +7,13 @@ import net.minecraftforge.fluids.capability.IFluidHandler
 import net.minecraftforge.items.IItemHandlerModifiable
 import net.minecraftforge.items.ItemHandlerHelper
 import net.minecraftforge.oredict.OreDictionary
+import org.ender_development.catalyx.api.v1.extensions.copyOf
 import org.ender_development.catalyx.core.recipes.chance.output.ChancedFluidOutput
 import org.ender_development.catalyx.core.recipes.chance.output.ChancedItemOutput
 import org.ender_development.catalyx.core.recipes.chance.output.ChancedOutputList
 import org.ender_development.catalyx.core.recipes.ingredients.RecipeInput
 import org.ender_development.catalyx.core.recipes.ingredients.RecipeInputCache
 import org.ender_development.catalyx.core.utils.IItemStackHash
-import org.ender_development.catalyx.core.utils.extensions.copyOf
 import org.ender_development.catalyx.modules.integration.groovyscript.ModuleGroovyScript
 
 class Recipe (
