+++ src/main/kotlin/org/ender_development/catalyx/core/recipes/validation/Validator.kt
@@ -2,7 +2,7 @@ package org.ender_development.catalyx.core.recipes.validation
 
 import org.apache.logging.log4j.Logger
 import org.ender_development.catalyx.Catalyx
-import org.ender_development.catalyx.core.utils.extensions.validateWith
+import org.ender_development.catalyx.api.v1.extensions.validateWith
 import org.ender_development.catalyx.core.utils.validation.ValidationError
 
 class Validator {
