+++ src/main/kotlin/org/ender_development/catalyx/core/tiles/TESRTile.kt
@@ -4,14 +4,14 @@ import net.minecraft.client.Minecraft
 import net.minecraft.util.EnumFacing
 import net.minecraftforge.fml.relauncher.Side
 import net.minecraftforge.fml.relauncher.SideOnly
+import org.ender_development.catalyx.api.v1.extensions.relativeDirectionTo
+import org.ender_development.catalyx.api.v1.extensions.withAlpha
 import org.ender_development.catalyx.core.ICatalyxMod
 import org.ender_development.catalyx.core.client.tesr.AbstractTESRenderer
 import org.ender_development.catalyx.core.client.tesr.HudInfoRenderer
 import org.ender_development.catalyx.core.tiles.helper.HudInfoLine
 import org.ender_development.catalyx.core.tiles.helper.IHudInfoProvider
 import org.ender_development.catalyx.core.tiles.helper.ITESRTile
-import org.ender_development.catalyx.core.utils.extensions.relativeDirectionTo
-import org.ender_development.catalyx.core.utils.extensions.withAlpha
 import java.awt.Color
 
 open class TESRTile(mod: ICatalyxMod) : BaseTile(mod), ITESRTile, IHudInfoProvider {
