--- src/main/kotlin/org/ender_development/catalyx/core/utils/extensions/Int.kt
@@ -1,4 +1,0 @@
-package org.ender_development.catalyx.core.utils.extensions
-
-val Int.plural
-	inline get() = if(this == 1) "" else "s"
