+++ src/main/kotlin/org/ender_development/catalyx/modules/integration/IntegrationModule.kt
@@ -1,8 +1,8 @@
 package org.ender_development.catalyx.modules.integration
 
+import org.ender_development.catalyx.api.v1.extensions.subLogger
 import org.ender_development.catalyx.api.v1.modules.annotations.CatalyxModule
 import org.ender_development.catalyx.core.Reference
-import org.ender_development.catalyx.core.utils.extensions.subLogger
 import org.ender_development.catalyx.modules.CatalyxInternalModuleContainer
 import org.ender_development.catalyx.modules.CatalyxModuleBase
