+++ src/main/kotlin/org/ender_development/catalyx/modules/integration/groovyscript/ModuleGroovyScript.kt
@@ -4,11 +4,11 @@ import com.cleanroommc.groovyscript.GroovyScript
 import com.cleanroommc.groovyscript.api.GroovyPlugin
 import com.cleanroommc.groovyscript.compat.mods.GroovyContainer
 import net.minecraftforge.fml.common.Optional
+import org.ender_development.catalyx.api.v1.extensions.subLogger
 import org.ender_development.catalyx.api.v1.modules.annotations.CatalyxModule
 import org.ender_development.catalyx.core.Reference
 import org.ender_development.catalyx.core.module.ModuleManager
 import org.ender_development.catalyx.core.utils.Mods
-import org.ender_development.catalyx.core.utils.extensions.subLogger
 import org.ender_development.catalyx.modules.CatalyxInternalModuleContainer
 import org.ender_development.catalyx.modules.integration.IntegrationModule
