+++ src/main/kotlin/org/ender_development/catalyx/modules/integration/top/FluidTileProvider.kt
@@ -5,9 +5,9 @@ import mcjty.theoneprobe.apiimpl.styles.ProgressStyle
 import net.minecraft.block.state.IBlockState
 import net.minecraft.entity.player.EntityPlayer
 import net.minecraft.world.World
+import org.ender_development.catalyx.api.v1.extensions.getRealColor
 import org.ender_development.catalyx.core.Reference
 import org.ender_development.catalyx.core.tiles.helper.IFluidTile
-import org.ender_development.catalyx.core.utils.extensions.getRealColor
 import java.awt.Color
 
 internal class FluidTileProvider : IProbeInfoProvider {
