+++ src/main/kotlin/org/ender_development/catalyx/modules/integration/top/ModuleTheOneProbe.kt
@@ -2,10 +2,10 @@ package org.ender_development.catalyx.modules.integration.top
 
 import mcjty.theoneprobe.TheOneProbe
 import net.minecraftforge.fml.common.event.FMLInitializationEvent
+import org.ender_development.catalyx.api.v1.extensions.subLogger
 import org.ender_development.catalyx.api.v1.modules.annotations.CatalyxModule
 import org.ender_development.catalyx.core.Reference
 import org.ender_development.catalyx.core.utils.Mods
-import org.ender_development.catalyx.core.utils.extensions.subLogger
 import org.ender_development.catalyx.modules.CatalyxInternalModuleContainer
 import org.ender_development.catalyx.modules.integration.IntegrationModule
