+++ src/main/kotlin/org/ender_development/catalyx/modules/test/DevTestModule.kt
@@ -5,6 +5,7 @@ import net.minecraft.util.text.TextComponentString
 import net.minecraftforge.client.event.ClientChatEvent
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent
 import org.ender_development.catalyx.Catalyx
+import org.ender_development.catalyx.api.v1.extensions.subLogger
 import org.ender_development.catalyx.api.v1.modules.annotations.CatalyxModule
 import org.ender_development.catalyx.core.Reference
 import org.ender_development.catalyx.core.blocks.IOTileBlock
@@ -13,7 +14,6 @@ import org.ender_development.catalyx.core.blocks.multiblock.parts.CornerBlock
 import org.ender_development.catalyx.core.blocks.multiblock.parts.SideBlock
 import org.ender_development.catalyx.core.client.AreaHighlighter
 import org.ender_development.catalyx.core.utils.SideUtils
-import org.ender_development.catalyx.core.utils.extensions.subLogger
 import org.ender_development.catalyx.modules.CatalyxInternalModuleContainer
 import org.ender_development.catalyx.modules.CatalyxModuleBase
